OK, I take back what I said, xcode is pretty cool. She lives again!

+++ PTC_xcode/PTC_xcode/city/Road.h
@@ -39,7 +39,7 @@ public:
 	void setEndIntersection(Intersection* endNode);
 	Intersection* getStartIntersection();
 	Intersection* getEndIntersection();
-	int getRoadID();
+	long int getRoadID();
 	int* getElevData();
 	int* getSpeedData();
 	int getBoundsID();
